3 Minigames to Upload (Asteroids, Noughts and Crosses, and Pong)
Started last Sunday when the internet went out, 3 days later with no internet and here we are.

A bunch of changes can be made such as expandable options for Noughts & Crosses (N&C) and Pong. With N&C I might add above 3 width and height options. For Pong difficulty levels for the AI player would be a good addition. Asteroids might be good as an online multiplayer, aside from the local 2 player already installed connecting with people from 2 different computers would be good.