Android API level 36 Update
Updated Android API to level 36.
Made a new algorithm for calculating a win state on the diagonals. The new one will now take half the board and calculate a win based on being past this halfway point. This change was made because the game was kind of pointless at higher levels. The game now is less gimmicky and higher levels now offer a new game that is almost different from 3x3 noughts and crosses. This is a very exciting change. The diagonal scoring system is costly though necessary as it adds a wanted complexity to the game that turns the board into a different game with every size increase.

As put in some future update notes which can found in the noughts and crosses edit files. The game now has 3 new updates that can be done past this point:
1) Noughts and crosses within noughts and crosses. A game within a game, would be silly past 5x5 but still possible.
2) a formula to increase or decrease this diagonal win variable. Either automatic or changeable in the UI. Though this would cause confusion at an already somewhat confusing game.
3) A square it button. A button that when pressed will take the largest number and make the other number equal to said number for when the game is restarted. This will reduce time adjusting to the desired settings.